refactor: 객체를 활용해 자동차 경주의 결과들을 보관하는 기능 구현
